chore: move inmemhttp from firecracker provisioner to internal/pkg/
To be reused in qemu provisioner later on.
